Add --only to `mb service *` and `mb environment configure`
Allow filtering nodes when running service commands (and also on environment configure).
This uses existing NodeFilter functionality.

Using a node filter on `mb environment configure` is considered an anti-pattern, but
there may be times it is necessary. The option is hidden for that command, and is only
intended to be used in unusual situations (usually driven by automation).